+++ packages/cqrs/CHANGELOG.md
@@ -1,4 +1,13 @@
+# 10.0.1
+
+- **Breaking:** Rename `forbiddenAccess` to `authorization` in error enums.
+- Update error enums docs.
+- Fix `CqrsMiddleware` not being exported from the package.
+- Move `hasError` and `hasErrorForProperty` from `CommandResponse` to `CommandFailure`.
+
 # 10.0.0
+
+**Retracted.**
 
 - **Breaking:** Fundamentaly change overall `Cqrs` API making it no-throw guarantee.
 - **Breaking:** Make `Cqrs.get`, `Cqrs.run` and `Cqrs.perform` return result data in form of `QueryResult`, `CommandResult` and `OperationResult` respectively.
@@ -6,9 +15,9 @@
 - Add middleware mechanism in form of `CqrsMiddleware` intended to use in processing result from queries, commands and operations.
 - **Breaking:** Remove `CqrsException`.
 - **Breaking:** Rename previous `CommandResult` to `CommandResponse` and make it package private.
-- Mark the `CqrsMethod` as `sealed`.
-- **Breaking:** Make `ValidationError` extend `Equatable` from `equatable` package.
-- **Breaking:** Add `equatable` (`^2.0.5`) and `logging` (`^1.2.0`) dependencies.
+- **Breaking:** Mark the `CqrsMethod` as `sealed`.
+- Make `ValidationError` extend `Equatable` from `equatable` package.
+- Add `equatable` (`^2.0.5`) and `logging` (`^1.2.0`) dependencies.
 
 # 9.0.0